+++ sdks/node/src/server.ts
@@ -2,6 +2,7 @@ export * from './integration.js';
 export * from './nest.js';
 export * from './tasks/adapters.js';
 export * from './tasks/http.js';
+export * from './tasks/node-response.js';
 export * from './tasks/sse.js';
 export * from './tasks/websocket.js';
 export * from './tasks/result.js';